submit C.cpp to 'C - 3 Steps' (code-festival-2017-qualb) [C++14 (GCC 5.4.1)]

+++ 2017/1008_CF2017-qualB/C.cpp
@@ -35,7 +35,7 @@ typedef tuple<int, int, int> state; // now, from, clr
 typedef tuple<int, int> cond; // now, depth
 
 ll N;
-int M;
+ll M;
 vector<int> V[100010];
 
 int main () {
